MERGES 12910
git-svn-id: svn://openspcoop.org/gov4j/openspcoop2/branches/3.0/3.0.dev@12913 f7a281a6-e321-0410-a5c7-cb86b8957a31

+++ tools/web_interfaces/control_station/src/org/openspcoop2/web/ctrlstat/core/ControlStationJMSCore.java
@@ -64,7 +64,7 @@ public class ControlStationJMSCore {
 
 		// Estraggo i dati dall'Operazione da smistare, che tanto mi servono
 		// per settare la StringProperty
-		int idTable = operazioneDaSmistare.getIDTable();
+		long idTable = operazioneDaSmistare.getIDTable();
 		Operazione operazione = operazioneDaSmistare.getOperazione();
 		String pdd = operazioneDaSmistare.getPdd();
 		TipoOggettoDaSmistare oggettoDaSmistare = operazioneDaSmistare.getOggetto();
